requirements.txt
alembic==1.17.1
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
bcrypt==4.0.1
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
Jinja2==3.1.6
python-jose>=3.3.0
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2-binary==2.9.11
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
